Reorganize files
- Created SharedPreferencesHelper to handle
  all shared preferences calls
