+++ app/src/main/java/com/foodmap/app/ui/mainfragment/ProfileFragment.java
@@ -1,8 +1,6 @@
-package com.foodmap.app.ui.main;
+package com.foodmap.app.ui.mainfragment;
 
-import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
@@ -16,8 +14,9 @@ import android.widget.Button;
 import com.auth0.android.authentication.AuthenticationException;
 import com.auth0.android.callback.Callback;
 import com.foodmap.app.R;
-import com.foodmap.app.model.Auth0Manager;
+import com.foodmap.app.model.auth.Auth0Manager;
 import com.foodmap.app.ui.LoginActivity;
+import com.foodmap.app.ui.SharedPreferencesHelper;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -56,7 +55,7 @@ public class ProfileFragment extends Fragment {
 
                 @Override
                 public void onSuccess(Void result) {
-                    clearSharedPreferences();
+                    SharedPreferencesHelper.clearUserPrefs(getContext());
 
                     // close main activity
                     Intent intent = new Intent(getActivity(), LoginActivity.class);
@@ -72,11 +71,6 @@ public class ProfileFragment extends Fragment {
         });
 
         return view;
-    }
-
-    private void clearSharedPreferences(){
-        SharedPreferences sharedPreferences = getContext().getSharedPreferences("SHARED_PREFS_USER", Context.MODE_PRIVATE);
-        sharedPreferences.edit().clear().commit();
     }
 
 }
