+++ app/src/main/res/layout/fragment_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".ui.main.ListFragment">
+    tools:context=".ui.mainfragment.ListFragment">
 
     <SearchView
         android:id="@+id/searchView"
